// src/externals/omdbapi/interfaces/omdbapi-response.interface.ts
import { OmdbapiRatingInterface } from './omdbapi-rating.interface'

export interface OmdbapiResponseInterface {
    Title: string
    Year: string
    Rated: string
    Released: string
    Runtime: string
    Genre: string
    Director: string
    Writer: string
    Actors: string
    Plot: string
    Language: string
    Country: string
    Awards: string
    Poster: string
    Ratings: OmdbapiRatingInterface[]
    Metascore: string
    imdbRating: string
    imdbVotes: string
    imdbID: string
    Type: string
    DVD: string
    BoxOffice: string
    Production: string
    Website: string
    Response: string
}